Relax RSpec dependency to fix CI build
As we don't specify a version of Rake, builds are now breaking due to
the removal of last_comment in Rake 11 and its usage in our outdated
version of RSpec.

We could fix this by specifying a version constraint for all
development dependencies but we can also relax our constraint on RSpec
so that we pull in a more recent version without this bug.